Allow for zlib compression on permacache values
This patch adds the ability, but doesn't turn it on yet.

This should reduce memtable requirements on Cassandra and have some other
benefits like reducing bandwidth requirements for giant keys. It will cost a
little bit of CPU and memory. CPU is tracked via timers.

It also adds the ability to use JSON format for some keys which can reduce
deserialisation time