Timeouts: Fix the timeout days remaining buffer
A one hour buffer was added so that timeouts wouldn't be seen as
permanent if TryLater hadn't processed it immediately when it expired.
That added an hour to the beginning of the timeout which increased the
display of the timeout by 1 day. This will just return 1 day if the
timeout expired but the TryLater queue hasn't been processed instead
of adding an hour.