Subreddit Rules: Remove XSS sinks
The initial rendering of the rules is done in mako, but all subsequent
rendering of the rules is done client-side. Using the wrong template
tags here creates an unnecessary vulnerability that could be used to
XSS attack a limited set of users (mods could potentially XSS attack
other mods of the same sub).