Link ID redirect: move into FrontController
To be able to do shortlink/id36 redirects, the user context has to be
set up to be able to check whether the user has access to view the
subreddit that the post is in. So this endpoint needs to be inside a
controller that's based on RedditController to make sure that info is
available.