events: Use baseplate events library
The primary goal of this is to standardize on the Baseplate base Event
class and remove AMQP from the events pipeline. The latter is done by
using Baseplate's EventQueue (which is based on POSIX queues local to
the box) and Baseplate's event publisher daemons which will take those
serialized events and ship them in batch to the collector.

The baseplate event library deliberately does not implement event field
truncation at this time as an effort to simplify. Metrics show that
since the events were boosted in size to 100 KiB, we have neither
truncated nor dropped an event. We'll have metrics here for oversized
events being dropped, and if it starts happening again we can bring back
truncation.