DO NOT OPENSOURCE: temp test of compose ratelimit
This starts checking a ratelimit on POST_compose, but doesn't actually
do any limiting yet. It will send an alert to me on Slack whenever
someone *would* have been ratelimited, so that I can keep an eye on this
for a couple of days to see if there's a large number of false positives
or some way that we need to be narrowing this down instead of just
naively ratelimiting like this.