Outbound Links: Add expiration to each link for validation
We're seeing a small-but-measurable number of expired clicks
coming through in our 1%-logged-in test. This is an attempt
to fix those and also is maybe a bit more long term robust.

This alters each outbound link generated to be an object, so that
we can pass back both the generated url and its expiration. By
passing its expiration explicitly, we gain a couple of benefits:

1. In the event this is cached, we know explicitly when it
expires. (technically, we could pull this out of the signature
in python, but that is more challenging in JS).

2. It avoids defining expiration deltas in more than one place, so
if we alter the expiration delta later, we can just change it once.