Scraper og:image check: handle no content attr
The thumbnail scraper is crashing on unusual og:image tags, when they
don't have a content= attribute. Example from the wild:

<meta property="og:image">

This just gets rid of the assumption that the attr will be there, so
that it can carry on finding a thumbnail as normal if an unusual tag is
encountered.