Gold: Use a more random string for create_gift_gold to avoid collisions
We're about to launch our native apps, which will have many folks
logging in for the first time and receiving some thank-you gold
as they log in.

Unfortunately, way back in the day `create_gift_gold` was created
in a way that is super collision unfriendly. If many folks log in
at the same time (very likely), they will collide on creating a
gift gold presently.

This increases the random string from 2 to 8 characters, to make
collisions within the same second unlikely. Transaction IDs are
allowed to be longer than this length, and I've tested in the shell
in production that a transaction ID of this length doesn't cause
any errors on insert.