Add helper for running Gunicorn workers under Einhorn
Einhorn is a language/protocol-agnostic socket and worker manager.
We're using it elsewhere for Baseplate services (where something
WSGI-specific wouldn't work on Thrift-based services) and its graceful
reload logic is more friendly than Gunicorn's. However, for non-gevent
WSGI, we still need something to parse HTTP and provide a WSGI
container. Gunicorn is excellent at this. This adds support for running
a paster application in a gunicorn worker under Einhorn.