Orangered emails: remove translation from templates
Translation does not work for emails, because the messages are drained from a queue where the user's language preference is not available. In addition, the old templates were not handling translation properly when interpolating variables. It did not seem worth it to fix the translation for this feature without a way to test it, and without translation working for emails in general