Allow stickies to be auto-hidden outside the SR
With the fix to .stickied values in 972bf0f, this is now preventing
stickies from being auto-hidden due to voting/score preferences while in
combined listings like /r/all.

This adds a check to only disable the auto-hiding while we're inside the
same subreddit as the sticky is posted in.