Simplify ratelimit.py's API
The fundamental operations that the user of the library cares about are
"check if I can take an action" and "record that I took, or am about to
take, an action".

The standlone functions provide the tools needed to build the above
operations. They do not have any policy beyond the translation of their
use case to the underlying key-value store operations, so they are
valuable in unusual cases. The new RateLimit uses them to make the
probably-common case of ratelimits pulled from g.live_config simpler. It
provides check() and record_usage() operations.

Also, add documentation and _multi variants and support rate limits with
greater than 1 day periods, despite that they are more likely to fail
due to cache resets.