Completely disable the pagecache
This is an experiment to see if the pagecache gives us any real benefit
given all the other layers of caching in r2. As the logged out
experience becomes more user-customized, this cache is going to be more
and more ineffective so it's good to see if we can just tear out the
layer now and save ourselves a lot of trouble.