Experiment: redirecting links and expandos
When a user clicks on a link on the listings page
they will be redirected to the comments page for
all links with media previews available. Clicking
the thumbnail from a listings page will open the
expando inline. Users can navigate to the original
content via the link on the comments page