Fix default thing button styling when banning.
By default, we render ban buttons on all things where applicable.  The inclusion of big_modbuttons makes this more complicated, and actually means when big_modbuttons are not rendered and the thing uses default styling and the thing has been reported, we don't show ban buttons at all.

This fixes that and also adds a report counter by default.