New top bar experiment
Experiment new_topbar has 2 variants, the first of which adds
a new top bar styled similar to the mobile website. The top bar
has categories from multireddits made by u/redditdefaults. Search and
the login/user info was moved into it as well. We hope to see if this
increases engagement. Subreddit style sheets are turned off as they cause
major issues to the appearance and functionality. We added the second
variant which does nothing except turn off stylesheets to help us
compensate for that in our analysis.