POST_subscribe: add param to skip default subs
When a user subscribes to a subreddit for the first time, we also
subscribe them to the full set of current default subreddits. This is
because users without customized subscriptions are effectively
"soft-subscribed" to the default set, and the first
subscribe/unsubscribe action converts these into "real" subscriptions.

This commit adds a new skip_initial_defaults param to the POST_subscribe
endpoint, which can be set to True while doing a user's initial
subscription action to skip over subscribing to the default set. So then
the user's first subscription will result in them being subscribed to
ONLY the specified subreddit(s), instead of in addition to the default set.

As part of this, the logic that sets a user's has_subscribed attribute
was moved into the individual subscription functions in Subreddit,
instead of only being set in subscribe_defaults(), since that will no
longer always be called on a user's initial subscription.