Use advocate for HTTP requests caused by clients
This should add a layer of protection against SSRF attacks by checking
that any requests we generate are not going to try to connect to any
local / private servers.

This also has the nice benefit of getting rid of a lot of custom code
because we can now take advantage of Requests's decoding etc.