Orangered emails: make user list unique to reduce queue items
The trylater queue has been backed up from our orangered
emails. Our hypothesis is that the user list that gets
passed into the message_notification_email function is
not unique, and was resulting in thousands of extra
lookups. This change forces the user list to be unique
to reduce our DB calls.