Add support for clientside websockets per session
This checks to see if a websocket for a certain url is open
in a current tab and if not open a new one. If a new websocket
is opened and the old one comes back to life (it was paused or
something), the old one will close itself.

A heartbeat (timestamp written in localStorage) notifies all tabs
that a websocket is currently running. When the heartbeat stops for
a time period (currently 15 seconds), the tabs will assume a websocket
was closed and will open a new one.

`websocketEvents` are the events that the websocket expects to receive
and will be handled in the tab that has the websocket open.
`websocketStorageEvents` are event handlers for when the
`websocketEvents` event is written to localStorage so that all tabs
are notified of the event.