EXPERIMENT: Read subreddit sidebars directly from Wiki
We're looking at doubling the maximum size of sidebars. A concern here
is that they are currently cached on the Subreddit object and so
increasing their size will increase the pressure on the caches (both
space and network traffic). We already store the sidebar separately in
the Wiki for revision history, so we can try reading from there directly
rather than using the cached version. If this works, we can stop caching
on the subreddit altogether.