Add whitelist to SSRF protections
Within the reddit VPC we resolve *.reddit.com to the internal IP of our
load balancers. This allows things like mobile web to resolve the API
without leaving the AWS private network. Unfortunately, this also means
that reddit.com links look like SSRF attempts to Advocate when
protecting the scraper. This adds a manual whitelist of internal IPs
that we are OK with the scraper contacting.