media.py: Add _GfycatScraper
Because we currently require scraping gifs to generate mp4 previews,
and it's not possible to determine the gif URL for any given gfycat
link, we need a special scraper that will hit their API to find the
proper URL to scrape.  When the gif source is found, this scraper
behaves like our normal _ImageScraper.  If the API request to find the
gif fails, or if the source gif is too large, the link will fall back
to being scraped by the _EmbedlyScraper.