Add comment warning about possible redirect loop
Mweb is about to begin understanding the mweb redirect
cookie. In doing so it introduces the possibility of a
redirect loop.

Since the default state of r2 is to redirect mobile users to
mweb and mweb will now try to redirect users to desktop if
they have the "no-redirect" cookie, not reading that cookie
would cause a redirect loop.