Start tracking activity for logged-out visitors
This starts writing to the activity service for logged-out visits, using
the LOID as the visitor ID. We then always read this, but only show it
if the feature is enabled. This will allow us to get an idea of the
numbers before making the proper front-end changes and turning this on
for everyone.