zookeeper: Add ReducedLiveList and IPNetworkLiveList
These classes are like LiveList (and use LiveList internally) but
they also store and watch a copy of the reduced data. This is useful
for cases where the map/reduce phase is slow and CPU intensive.
By storing the reduced data separately the map/reduce only needs
to be done by the process that's updating the list. All other processes
watch the reduced data node.