Add live comments for links that are eligible
Serverside: when a comment has been posted and is in a link that is
eligible for live comments, the comment will be added to a queue where
it will wait a few seconds so that it's had a chance to go through
automod and antispam processing. The queue processor will then render
the comment html and also a mod version of the comment (so that the
proper flatlist buttons are added) and broadcast it.

Clientside: when the new comment message has been received, it will
determine where the element should be placed and if that spot is in
the current viewport it will add the element and display it. If it
isn't currently in view, it will add the element but keep it hidden and
show the modal that the user can click to display all the new comments.