Hide modmail sent to new system from old system
In order to make sure that the test subreddits for the new modmail
system are actually using it, we want to only show new conversations in
the new system.

This adds a `new_modmail_enabled_at` attr to a subreddit that should be
set to the datetime that new modmail is enabled. Any new message threads
after that point will be hidden. If necessary for the subreddit to opt
back out of the test, setting this attr back to None will make all of
the modmail show up again.