Move new modmail creation into Message._new()
For integration purposes, we need to have modmail messages created in
both the new modmail system as well as the legacy messaging system.

Previously, if a modmail was sent via the legacy system (which all users
will continue to do), the corresponding conversation/message in the new
system was being created inside the API endpoint code. This commit moves
that process into Message._new(), so that internal code creating modmail
will go through this as well, instead of only calls coming in through
the external endpoints.