Add is_auto to ModmailConversation constructor
Previously, ModmailConversations with is_auto set to True (which puts
them in the "notifications" folder) were only created if the sender was
AutoModerator. This adds an argument to the constructor that will enable
us to create them in other cases as well. A related argument is added to
Message._new() to support creating a notification modmail via that
function as well.