Don't send 2nd mod invite message to new modmail
Currently, when a moderator is invited, two messages are sent out:

1. A message from the subreddit to the invited user (which does not
   show up in current modmail)
2. A message from the system user (/u/reddit) to the subreddit (does
   appear in current modmail)

The new modmail system is showing both of these, so we're going to stop
sending the second message from the system user and only have the first
one.

Once new modmail is active everywhere and is the only form of modmail,
this code can be updated a bit more to just fully eliminate that second
message so this special check isn't necessary.