[API-279] read_all_messages returns no HTML if raw_json=1 (#5791)
Currently, this endpoint always returns some HTML that says

```
<head>
 <title>202 Accepted</title>
</head>
<body>
 <h1>202 Accepted</h1> </body>
```

The gateway system expects to only receive json (or nothing) from r2 so this valid response will cause an error in the gateway.  Rather than adding in special cases for things like this, this change just updates the endpoint to respect the raw_json flag sent by the gateway.

It would probably be ideal to just update the endpoint to not use `abort` to send a 202 status code but there are apps out in the world using this endpoint so we may not want to change it in all cases.