Old modmail system: hide messages sent to new one
Previous attempt (3b50012f) did this hiding inside the builder, but
MessageBuilder doesn't handle this well and can result in small/empty
pages, and potential 403 errors if the last item on the page was hidden.

This commit switches to using a root message ID as the "threshold"
instead of a datetime, and uses that message ID to filter out modmail
trees earlier in the process, so that the builder doesn't need to handle
it.

Functions for enabling and disabling new modmail (with the cutoff being
the current last message in the subreddit's modmail) have been added to
Subreddit as well, to make this simpler.