Ads: assign default third element on SubredditsByPartialName tups
Makes the unpacking of the tups attribute from the returned record
from SubredditsByPartialName._byID agnostic to additional values.
This commit should be deployed in conjunction with ff576dd, which
adds an additional value to the aforementioned record.

Changes from this commit should be short-lived and will be
overwritten by the commit titled "Ads: check include_unadvertisable
when querying SubredditSelector". This commit only needs to live
for as long as it takes to update SubredditsByPartialName records
following the deployment of the commit titled "Ads: add
'unadvertisable' when running load_all_reddits".